fix: Check for editing `div` elements
This is just a quickfix to solve `trello` and similar websites that keep
the editor and preview in sync.
This commit checks if the target is an editing scope because if the website keeps the
editor and preview in sync, when I change one of them by adding the class name,
it gets to an infinite loop by changing both at the same time.